==> 1_code/r_notebook/2_harvestData.Rmd <==

```{r setup, include=FALSE, cache=FALSE}
#Set root directory to R project root
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

```{r eval=FALSE}
library(sf)
```

## CAS-FRI Harvest data {-}

### Import CAS-FRI polygons {-}
```{r eval=FALSE}
fgdb <- "../Data/CASFRI/Alberta/alberta_geo.gdb"
fc_list <- ogrListLayers(fgdb)

casfri<-st_read("../Data/CASFRI/alberta_geo.gdb", layer = "alberta_geo")

casfri<-casfri%>%
  dplyr::select(cas_id)
```


### Get disturbance polygons
```{r}
cas_dst<-read.csv("../Data/CASFRI/alberta_dst.csv")

#get only harvest polygons
cas_dst_2<-cas_dst%>%
  dplyr::filter(dist_1=="CO"| dist_1=="PC")

# Keep only CASFRI polygons with silverculture disturbance history
casfri_harvest<-casfri%>%
  inner_join(cas_dst_2, by="cas_id")

st_write(casfri_harvest, "0_data/manual/spatial/harvest/casfri_harvest.shp")
```

## ABMI harvests {-}
```{r}
abmi<-st_read("../Data/ABMI/HFI_2019_v2.gdb", layer="o18_HarvestAreas_HFI_2019")
abmi$uniqueID <- 1:nrow(abmi) 

####################################################
### Processes geometries

# standardize geometries
# check type of geometries
as.data.frame(table(st_geometry_type(abmi)))
#operations not supported on on mmultisurface geometries so cast as multipolygons
abmi = st_cast(abmi, "MULTIPOLYGON")
# check types
as.data.frame(table(st_geometry_type(abmi)))
st_write(abmi, "0_data/manual/spatial/harvest/o18_HarvestAreas_HFI_2019.shp")


# check geometry errors
valid<-st_is_valid(abmi, reason = TRUE)
valid<-as.data.frame(valid)
nonvalid<-valid%>%filter(valid!="Valid Geometry")

abmi_2<-st_buffer(abmi, dist=0)
# repair geometries
abmi_2<-st_make_valid(abmi)
# abmi_3<-st_buffer(abmi_2, 0.0)
#check
valid_2<-st_is_valid(abmi, reason = TRUE)
valid_2<-as.data.frame(valid_2)
nonvalid_2<-valid_2%>%filter(valid_2!="Valid Geometry")

####################################
# remove donut holes from polygons
####################################

library(nngeo)
abmi_3<-st_remove_holes(abmi_2)
valid_3<-st_is_valid(abmi_3, reason = TRUE)
valid_3<-as.data.frame(valid_3)
st_write(abmi_3, "0_data/manual/spatial/harvest/o18_HarvestAreas_HFI_2019_noHoles.shp")
"0_data/manual/spatial/harvest/o18_HarvestAreas_HFI_2019_noHoles.shp"
####################################
## create a 30 m negative buffer
####################################

# check units
st_crs(abmi_3)$units
#buffer
abmi_buff<-st_buffer(abmi_3, dist=-30)
# check geometries
valid_4<-st_is_valid(abmi_buff, reason = TRUE)
valid_4<-as.data.frame(valid_4)

as.data.frame(table(st_geometry_type(abmi_buff)))
abmi_buff = st_cast(abmi_buff, "MULTIPOLYGON")
st_write(abmi_buff, "0_data/manual/spatial/harvest/o18_HarvestAreas_HFI_2019_buff_neg30.shp")

####################################
## Filter to surveyed polygons
####################################
load("0_data/manual/spatial/bird/ss_xy.rData")
ss_xy <- st_transform(ss_xy, crs = 3400)

# surveyed<-st_join(abmi_buff, ss_xy)
# surveyed_2<-surveyed%>%
#   filter(!is.na(SS))%>%
#   # dplyr::select(-c(uniqueID, survey_year))%>%
#   mutate(uniqueID=1:nrow(.))
# #4489  
# 
# st_write(surveyed_2, "0_data/manual/spatial/harvest/o18_HarvestAreas_HFI_2019_buff_neg30_surveyed.shp")
# 
# surveyed_3<-st_simplify(surveyed_2, dTolerance = 5)
# st_write(surveyed_3, "0_data/manual/spatial/harvest/o18_HarvestAreas_HFI_2019_buff_neg30_surveyed_simp.shp")

surveyed<-st_filter(abmi_buff, ss_xy)
st_write(surveyed, "0_data/manual/spatial/harvest/o18_HarvestAreas_HFI_2019_buff_neg30_surveyed_20220315.shp")
surveyed_2<-st_simplify(surveyed, dTolerance = 5)
st_write(surveyed_2, "0_data/manual/spatial/harvest/o18_HarvestAreas_HFI_2019_buff_neg30_surveyed_simp_20220315.shp")

```


### Join with point count locations {-}

```{r eval=FALSE}
casfri_harvest<-st_read("0_data/manual/spatial/harvest/casfri_harvest.shp")
abmi_harvest<-st_read("0_data/manual/spatial/harvest/o18_HarvestAreas_HFI_2019_noHoles.shp")
load("0_data/manual/spatial/bird/ss_xy.rData")

casfri_harvest <- st_transform(casfri_harvest, crs = 3400)
ss_xy <- st_transform(ss_xy, crs = 3400)
```

#### Distance to edge {-}

Calculate distance of point-coints to nearest harvest edge.

```{r eval=FALSE}
# get the boundary of shape files
pb<-st_boundary(abmi_harvest)

#identify the nearest polygon to each point
nearest = st_nearest_feature(ss_xy, abmi_harvest)

 #Then use st_distance to get the element-wise distances between each element of points and the corresponding element of boundary polygons:
dist = st_distance(ss_xy, pb[nearest,], by_element=TRUE)

# join the ponts and nearest polygons
bd_xy_edge = cbind(ss_xy, st_drop_geometry(pb)[nearest,])

#Then add the distance values
bd_xy_edge$dist_to_edge = dist
bd_xy_edge<-bd_xy_edge%>%
  as.data.frame()%>%
  dplyr::select(survey_year, SS, HFI_ID, dist_to_edge)

save(bd_xy_edge, file ="2_pipeline/tmp/bd_xy_edge.rData")
```

#### Perimeter area ratio {-}

Calculate the perimeter area ratio () of harvest polygons.
```{r eval=FALSE}
abmi_harvest<-st_read("0_data/manual/spatial/harvest/o18_HarvestAreas_HFI_2019_noHoles.shp")

abmi_paRatio<-abmi_harvest%>%
  mutate(perimeter=st_length(st_cast(geometry, "MULTILINESTRING")))%>%
  mutate(area=st_area(geometry))%>%
  mutate(pa_ratio=perimeter/area)

save(abmi_paRatio, file ="2_pipeline/tmp/abmi_paRatio.rData")
```

#### Join {-}

```{r eval=FALSE}
ss_xy_harvest<-st_join(ss_xy, casfri_harvest)%>%
  st_join(abmi_paRatio)%>%
  dplyr::filter(!is.na(cas_id)|!is.na(HFI_ID))

bd_xy_edge2<-bd_xy_edge%>%dplyr::select(SS, dist_to_edge)%>%
distinct()

ss_xy_harvest_df<-ss_xy_harvest%>%
  dplyr::select(c("SS","survey_year", "cas_id", "dist_1", "dst_y_1", "dst_xt_p_1", "dist_2", "dst_y_2", , "dst_xt_p_2", "HFI_ID", "uniquID", "YEAR", "Mdfr_Yr", "perimeter", "area", "pa_ratio" ))%>%
  dplyr::mutate(dst_y_2 = replace(dst_y_2, dst_y_2== -1111, NA))%>%
  dplyr::mutate(dist_1_2_diff=dst_y_2-dst_y_1)%>%
  dplyr::rename(c(dist_yr_1=dst_y_1, dist_yr_2=dst_y_2, dist_ext_upper_1=dst_xt_p_1, dist_ext_upper_2=dst_xt_p_2))%>%
  as.data.frame()%>%
  # left_join(bd_xy_edge, by = c("SS", "survey_year", "HFI_ID"))%>%
  left_join(bd_xy_edge2, by = c("SS"))%>%# add distance to edge column
  dplyr::select(c("SS","survey_year", "dist_to_edge", "cas_id",  "dist_1", "dist_yr_1", "dist_ext_upper_1", "dist_2", "dist_yr_2", "dist_ext_upper_2", "dist_1_2_diff", "HFI_ID", "uniquID", "YEAR", "Mdfr_Yr", "perimeter", "area", "pa_ratio" ))%>%
  mutate(perimeter=as.numeric(perimeter))%>%
  mutate(area=as.numeric(area))%>%
  mutate(pa_ratio=perimeter/area)
#   dplyr::filter(survey_year>=YEAR|survey_year>=dst_y_1)
#            as.data.frame()
save(ss_xy_harvest_df, file=paste0("0_data/manual/spatial/ss_xy_harvest_df_", Sys.Date(), ".rData"))

nrow(as.data.frame(unique(ss_xy_harvest_df$SS)))
#3752         
```


### ABMI regen metrics {-}

```{r eval=FALSE}
abmi_regen<-st_read("../Data/ABMI/ABMI_HarvAreaSpecRegn2019_v1/ABMI_HarvestArea_SpectralRegeneration_2019_v1.shp")

load("0_data/manual/spatial/bird/ss_xy.rData")
ss_xy <- st_transform(ss_xy, crs = 3400)

ss_xy_abmiRegen_df<-as.data.frame(st_join(ss_xy, abmi_regen))%>%
  filter(!is.na(HFI_ID))

save(ss_xy_abmiRegen_df, file="2_pipeline/tmp/ss_xy_abmiRegen_df.rData")
```
